Refactor: client.py became too large/difficult to understand; .py files and their contained functions did not correspond to another
Solution: refactor client.py into own module; rename pagination to page_size where possible; refactor models module
