+++ src/aleph/sdk/client/__init__.py
@@ -1,0 +1,12 @@
+from .authenticated import AuthenticatedAlephClient, AuthenticatedUserSessionSync
+from .base import BaseAlephClient, BaseAuthenticatedAlephClient
+from .client import AlephClient, UserSessionSync
+
+__all__ = [
+    "BaseAlephClient",
+    "BaseAuthenticatedAlephClient",
+    "AlephClient",
+    "AuthenticatedAlephClient",
+    "UserSessionSync",
+    "AuthenticatedUserSessionSync",
+]
